+++ src/main/java/nom/bdezonia/zorbage/type/data/float32/real/Float32VectorMember.java
@@ -227,7 +227,7 @@ public final class Float32VectorMember
 	
 	@Override
 	public PrimitiveRepresentation preferredRepresentation() {
-		return PrimitiveRepresentation.DOUBLE;
+		return PrimitiveRepresentation.FLOAT;
 	}
 
 	@Override
